- Rename "hash" so that there is no ambiguity with std::hash. - Avoid a macro redefinition. - Make some signed/unsigned conversions explicit. - Fix a few unused variables. - Reduce the scope of a few variables. - Don't use extra class name qualificater in header files. - Initilialize a few variables. - Fix the case for some file names. - Remove some useless casts. - Cosmetics (remove empty lines at the end of some files, ...)
Patch by XhmikosR.

git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@3605 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/SubPic/SubPicQueueImpl.cpp
@@ -427,7 +427,7 @@ DWORD CSubPicQueue::ThreadProc()
 		if(SUCCEEDED(GetSubPicProvider(&pSubPicProvider)) && pSubPicProvider
 				&& SUCCEEDED(pSubPicProvider->Lock())) {
 			for(POSITION pos = pSubPicProvider->GetStartPosition(rtNow, fps);
-					pos && !m_fBreakBuffering && GetQueueCount() < (size_t)nMaxSubPic;
+					pos && !m_fBreakBuffering && GetQueueCount() < nMaxSubPic;
 					pos = pSubPicProvider->GetNext(pos)) {
 				REFERENCE_TIME rtStart = pSubPicProvider->GetStart(pos, fps);
 				REFERENCE_TIME rtStop = pSubPicProvider->GetStop(pos, fps);
@@ -509,7 +509,7 @@ DWORD CSubPicQueue::ThreadProc()
 						AppendQueue(pDynamic);
 						bAgain = true;
 
-						if (GetQueueCount() >= (size_t)nMaxSubPic) {
+						if (GetQueueCount() >= nMaxSubPic) {
 							break;
 						}
 					}
